Remove form to enter API KEY for Open AI
When uploading PDF files, the environment variable OPENAI_API_KEY must be set, but when asking a question from the "Ask" form, the value can be obtained from the form of Open AI's API KEY.
These behaviors are inconsistent and confusing to end users.
Therefore, the form to enter Open AI's API KEY has been removed.

The reason for not using the value of Open AI's API KEY form when uploading PDF files is that the post request from the screen and the PDF file upload process are different.
In order to use the form value of API KEY at the time of PDF file upload, it is necessary to send the API KEY in some way at the time of file upload.

modified:   app/main.py
modified:   app/services/summary_service.py